+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/run/TyeCommandLineState.kt
@@ -25,6 +25,29 @@ open class TyeCommandLineState(
             ?: throw ExecutionException("Path argument not specified.")
         val workingDirectory = runConfig.pathArgument?.parent?.path
 
+        val arguments = buildArguments(pathArgument)
+
+        return createHandler(workingDirectory, tyeToolPath, arguments)
+    }
+
+    private fun createHandler(
+        workingDirectory: String?,
+        tyeToolPath: String,
+        arguments: MutableList<String>
+    ): OSProcessHandler {
+        val commandLine = GeneralCommandLine()
+            .withParentEnvironmentType(GeneralCommandLine.ParentEnvironmentType.CONSOLE)
+            .withWorkDirectory(workingDirectory)
+            .withExePath(tyeToolPath)
+            .withParameters(arguments)
+
+        val handler = OSProcessHandler(commandLine)
+        handler.startNotify()
+        ProcessTerminatedListener.attach(handler, environment.project)
+        return handler
+    }
+
+    private fun buildArguments(pathArgument: String): MutableList<String> {
         val arguments = mutableListOf<String>()
         arguments.add("run")
 
@@ -50,17 +73,35 @@ open class TyeCommandLineState(
             arguments.add(runConfig.verbosityArgument)
         }
 
+        if (!runConfig.tagsArgument.isNullOrBlank()) {
+            arguments.add("--tags")
+            arguments.add(runConfig.tagsArgument!!)
+        }
+
+        if (runConfig.logsProvider != LogsProvider.NONE) {
+            arguments.add("--logs")
+
+            var logsProviderArg = runConfig.logsProvider.argumentName
+            if (runConfig.logsProvider.isProviderUrlEnabled() && !runConfig.logsProviderUrl.isNullOrBlank()) {
+                logsProviderArg += "=${runConfig.logsProviderUrl}"
+            }
+
+            arguments.add(logsProviderArg)
+        }
+
+        if (runConfig.tracesProvider != TracesProvider.NONE) {
+            arguments.add("--dtrace")
+
+            var tracesProviderArg = runConfig.tracesProvider.argumentName
+            if (runConfig.tracesProvider.isProviderUrlEnabled() && !runConfig.tracesProviderUrl.isNullOrBlank()) {
+                tracesProviderArg += "=${runConfig.tracesProviderUrl}"
+            }
+
+            arguments.add(tracesProviderArg)
+        }
+
         arguments.add(pathArgument)
 
-        val commandLine = GeneralCommandLine()
-            .withParentEnvironmentType(GeneralCommandLine.ParentEnvironmentType.CONSOLE)
-            .withWorkDirectory(workingDirectory)
-            .withExePath(tyeToolPath)
-            .withParameters(arguments)
-
-        val handler = OSProcessHandler(commandLine)
-        handler.startNotify()
-        ProcessTerminatedListener.attach(handler, environment.project)
-        return handler
+        return arguments
     }
 }