+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # tye-plugin Changelog
 
 ## [Unreleased]
+### Added
+- New options in the run configuration to send logs, traces or specify tags.
 
 ## [0.0.2]
 ### Added